+++ assets/js/base/components/cart-checkout/index.js
@@ -1,5 +1,4 @@
 export { default as AddressForm } from './address-form';
-export { default as CheckoutForm } from './form';
 export { default as FormStep } from './form-step';
 export { default as OrderSummary } from './order-summary';
 export { default as PlaceOrderButton } from './place-order-button';
