+++ assets/js/blocks/cart-checkout/checkout/block.js
@@ -2,44 +2,21 @@
  * External dependencies
  */
 import classnames from 'classnames';
-import { useMemo, useEffect } from '@wordpress/element';
+import { useEffect } from '@wordpress/element';
 import { __ } from '@wordpress/i18n';
 import {
-	AddressForm,
-	FormStep,
 	PlaceOrderButton,
 	Policies,
 	ReturnToCartButton,
-	ShippingRatesControl,
 } from '@woocommerce/base-components/cart-checkout';
-import { Form } from '@woocommerce/base-components/form';
-import { ValidatedTextInput } from '@woocommerce/base-components/text-input';
-import CheckboxControl from '@woocommerce/base-components/checkbox-control';
-import {
-	getCurrencyFromPriceResponse,
-	getShippingRatesPackageCount,
-	getShippingRatesRateCount,
-} from '@woocommerce/base-utils';
-import FormattedMonetaryAmount from '@woocommerce/base-components/formatted-monetary-amount';
 import {
 	CheckoutProvider,
 	useCheckoutContext,
 	useEditorContext,
-	useShippingDataContext,
 	useValidationContext,
-	StoreNoticesProvider,
 } from '@woocommerce/base-context';
-import {
-	useStoreCart,
-	usePaymentMethods,
-	useStoreNotices,
-	useCheckoutAddress,
-} from '@woocommerce/base-hooks';
-import {
-	CheckoutExpressPayment,
-	PaymentMethods,
-} from '@woocommerce/base-components/payment-methods';
-import { decodeEntities } from '@wordpress/html-entities';
+import { useStoreCart, useStoreNotices } from '@woocommerce/base-hooks';
+import { CheckoutExpressPayment } from '@woocommerce/base-components/payment-methods';
 import {
 	Sidebar,
 	SidebarLayout,
@@ -47,19 +24,14 @@ import {
 } from '@woocommerce/base-components/sidebar-layout';
 import { getSetting } from '@woocommerce/settings';
 import withScrollToTop from '@woocommerce/base-hocs/with-scroll-to-top';
-import {
-	CHECKOUT_SHOW_LOGIN_REMINDER,
-	CHECKOUT_ALLOWS_GUEST,
-	DISPLAY_CART_PRICES_INCLUDING_TAX,
-} from '@woocommerce/block-settings';
+import { CHECKOUT_ALLOWS_GUEST } from '@woocommerce/block-settings';
 
 /**
  * Internal dependencies
  */
+import CheckoutForm from './form';
 import CheckoutSidebar from './sidebar';
 import CheckoutOrderError from './checkout-order-error';
-import CheckoutOrderNotes from './checkout-order-notes';
-import NoShippingPlaceholder from './no-shipping-placeholder';
 import './style.scss';
 
 /**
@@ -74,29 +46,6 @@ const Block = ( props ) => {
 			<Checkout { ...props } />
 		</CheckoutProvider>
 	);
-};
-
-/**
- * Renders a shipping rate control option.
- *
- * @param {Object} option Shipping Rate.
- */
-const renderShippingRatesControlOption = ( option ) => {
-	const priceWithTaxes = DISPLAY_CART_PRICES_INCLUDING_TAX
-		? parseInt( option.price, 10 ) + parseInt( option.taxes, 10 )
-		: parseInt( option.price, 10 );
-	return {
-		label: decodeEntities( option.name ),
-		value: option.rate_id,
-		description: decodeEntities( option.description ),
-		secondaryLabel: (
-			<FormattedMonetaryAmount
-				currency={ getCurrencyFromPriceResponse( option ) }
-				value={ priceWithTaxes }
-			/>
-		),
-		secondaryDescription: decodeEntities( option.delivery_time ),
-	};
 };
 
 /**
@@ -117,49 +66,13 @@ const Checkout = ( { attributes, scrollToTop } ) => {
 		hasOrder,
 		hasError: checkoutHasError,
 		isIdle: checkoutIsIdle,
-		isProcessing: checkoutIsProcessing,
 		customerId,
-		onSubmit,
-		orderNotes,
-		dispatchActions,
 	} = useCheckoutContext();
-	const { setOrderNotes } = dispatchActions;
 	const {
 		hasValidationErrors,
 		showAllValidationErrors,
 	} = useValidationContext();
-	const {
-		shippingRates,
-		shippingRatesLoading,
-		needsShipping,
-	} = useShippingDataContext();
-	const { paymentMethods } = usePaymentMethods();
 	const { hasNoticesOfType } = useStoreNotices();
-	const {
-		defaultAddressFields,
-		shippingFields,
-		setShippingFields,
-		billingFields,
-		setBillingFields,
-		setEmail,
-		setPhone,
-		shippingAsBilling,
-		setShippingAsBilling,
-		showBillingFields,
-	} = useCheckoutAddress();
-	const addressFieldsConfig = useMemo( () => {
-		return {
-			company: {
-				...defaultAddressFields.company,
-				hidden: ! attributes.showCompanyField,
-				required: attributes.requireCompanyField,
-			},
-			address_2: {
-				...defaultAddressFields.address_2,
-				hidden: ! attributes.showApartmentField,
-			},
-		};
-	}, [ defaultAddressFields, attributes ] );
 
 	const hasErrorsToDisplay =
 		checkoutIsIdle &&
@@ -198,20 +111,6 @@ const Checkout = ( { attributes, scrollToTop } ) => {
 		);
 	}
 
-	const loginPrompt = () =>
-		CHECKOUT_SHOW_LOGIN_REMINDER &&
-		! customerId && (
-			<>
-				{ __(
-					'Already have an account? ',
-					'woo-gutenberg-products-block'
-				) }
-				<a href={ loginToCheckoutUrl }>
-					{ __( 'Log in.', 'woo-gutenberg-products-block' ) }
-				</a>
-			</>
-		);
-
 	const checkoutClassName = classnames( 'wc-block-checkout', {
 		'has-dark-controls': attributes.hasDarkControls,
 	} );
@@ -220,207 +119,15 @@ const Checkout = ( { attributes, scrollToTop } ) => {
 			<SidebarLayout className={ checkoutClassName }>
 				<Main className="wc-block-checkout__main">
 					{ cartNeedsPayment && <CheckoutExpressPayment /> }
-					<Form
-						className="wc-block-components-checkout-form"
-						onSubmit={ onSubmit }
-					>
-						<FormStep
-							id="contact-fields"
-							disabled={ checkoutIsProcessing }
-							className="wc-block-checkout__contact-fields"
-							title={ __(
-								'Contact information',
-								'woo-gutenberg-products-block'
-							) }
-							description={ __(
-								"We'll use this email to send you details and updates about your order.",
-								'woo-gutenberg-products-block'
-							) }
-							stepHeadingContent={ loginPrompt }
-						>
-							<ValidatedTextInput
-								id="email"
-								type="email"
-								label={ __(
-									'Email address',
-									'woo-gutenberg-products-block'
-								) }
-								value={ billingFields.email }
-								autoComplete="email"
-								onChange={ setEmail }
-								required={ true }
-							/>
-						</FormStep>
-						{ needsShipping && (
-							<FormStep
-								id="shipping-fields"
-								disabled={ checkoutIsProcessing }
-								className="wc-block-checkout__shipping-fields"
-								title={ __(
-									'Shipping address',
-									'woo-gutenberg-products-block'
-								) }
-								description={ __(
-									'Enter the physical address where you want us to deliver your order.',
-									'woo-gutenberg-products-block'
-								) }
-							>
-								<AddressForm
-									id="shipping"
-									onChange={ setShippingFields }
-									values={ shippingFields }
-									fields={ Object.keys(
-										defaultAddressFields
-									) }
-									fieldConfig={ addressFieldsConfig }
-								/>
-								{ attributes.showPhoneField && (
-									<ValidatedTextInput
-										id="phone"
-										type="tel"
-										label={
-											attributes.requirePhoneField
-												? __(
-														'Phone',
-														'woo-gutenberg-products-block'
-												  )
-												: __(
-														'Phone (optional)',
-														'woo-gutenberg-products-block'
-												  )
-										}
-										value={ billingFields.phone }
-										autoComplete="tel"
-										onChange={ setPhone }
-										required={
-											attributes.requirePhoneField
-										}
-									/>
-								) }
-								<CheckboxControl
-									className="wc-block-checkout__use-address-for-billing"
-									label={ __(
-										'Use same address for billing',
-										'woo-gutenberg-products-block'
-									) }
-									checked={ shippingAsBilling }
-									onChange={ ( isChecked ) =>
-										setShippingAsBilling( isChecked )
-									}
-								/>
-							</FormStep>
-						) }
-						{ showBillingFields && (
-							<FormStep
-								id="billing-fields"
-								disabled={ checkoutIsProcessing }
-								className="wc-block-checkout__billing-fields"
-								title={ __(
-									'Billing address',
-									'woo-gutenberg-products-block'
-								) }
-								description={ __(
-									'Enter the address that matches your card or payment method.',
-									'woo-gutenberg-products-block'
-								) }
-							>
-								<AddressForm
-									id="billing"
-									onChange={ setBillingFields }
-									type="billing"
-									values={ billingFields }
-									fields={ Object.keys(
-										defaultAddressFields
-									) }
-									fieldConfig={ addressFieldsConfig }
-								/>
-							</FormStep>
-						) }
-						{ needsShipping && (
-							<FormStep
-								id="shipping-option"
-								disabled={ checkoutIsProcessing }
-								className="wc-block-checkout__shipping-option"
-								title={ __(
-									'Shipping options',
-									'woo-gutenberg-products-block'
-								) }
-								description={
-									getShippingRatesRateCount( shippingRates ) >
-									1
-										? __(
-												'Select shipping options below.',
-												'woo-gutenberg-products-block'
-										  )
-										: ''
-								}
-							>
-								{ isEditor &&
-								! getShippingRatesPackageCount(
-									shippingRates
-								) ? (
-									<NoShippingPlaceholder />
-								) : (
-									<ShippingRatesControl
-										noResultsMessage={ __(
-											'There are no shipping options available. Please ensure that your address has been entered correctly, or contact us if you need any help.',
-											'woo-gutenberg-products-block'
-										) }
-										renderOption={
-											renderShippingRatesControlOption
-										}
-										shippingRates={ shippingRates }
-										shippingRatesLoading={
-											shippingRatesLoading
-										}
-									/>
-								) }
-							</FormStep>
-						) }
-						{ cartNeedsPayment && (
-							<FormStep
-								id="payment-method"
-								disabled={ checkoutIsProcessing }
-								className="wc-block-checkout__payment-method"
-								title={ __(
-									'Payment method',
-									'woo-gutenberg-products-block'
-								) }
-								description={
-									Object.keys( paymentMethods ).length > 1
-										? __(
-												'Select a payment method below.',
-												'woo-gutenberg-products-block'
-										  )
-										: ''
-								}
-							>
-								<StoreNoticesProvider context="wc/payment-area">
-									<PaymentMethods />
-								</StoreNoticesProvider>
-							</FormStep>
-						) }
-						{ attributes.showOrderNotes && (
-							<FormStep id="order-notes" showStepNumber={ false }>
-								<CheckoutOrderNotes
-									disabled={ checkoutIsProcessing }
-									onChange={ setOrderNotes }
-									placeholder={
-										needsShipping
-											? __(
-													'Notes about your order, e.g. special notes for delivery.',
-													'woo-gutenberg-products-block'
-											  )
-											: __(
-													'Notes about your order.',
-													'woo-gutenberg-products-block'
-											  )
-									}
-									value={ orderNotes }
-								/>
-							</FormStep>
-						) }
-					</Form>
+					<CheckoutForm
+						loginToCheckoutUrl={ loginToCheckoutUrl }
+						showApartmentField={ attributes.showApartmentField }
+						showCompanyField={ attributes.showCompanyField }
+						showOrderNotes={ attributes.showOrderNotes }
+						showPhoneField={ attributes.showPhoneField }
+						requireCompanyField={ attributes.requireCompanyField }
+						requirePhoneField={ attributes.requirePhoneField }
+					/>
 					<div className="wc-block-checkout__actions">
 						{ attributes.showReturnToCart && (
 							<ReturnToCartButton
